requirements.txt
flask
flask-pymongo
python-jose 
requests 
cachetools
flask-socketio
gevent
gevent-websocket
flask-cors
redis